뉴스  기사 내 í내의 투표

+++ src/main/java/com/tave/alarmissue/news/converter/NewsConverter.java
@@ -14,7 +14,7 @@ public class NewsConverter {
                 .source(news.getSource())
                 .date(news.getDate())
                 .thema(news.getThema())
-                .sentiment(news.getSentiment() != null ? news.getSentiment().doubleValue() : 0.0)
+                .sentiment(news.getSentiment() != null ? news.getSentiment() : 0.0)
                 .sentiment(news.getSentiment())
                 .view(news.getView())
                 .imageUrl(news.getImageUrl())
